fix: overflow tests, negation overflow detection, lexer i128 widening
Fixes two bugs found by Ornith during test authoring:

1. Unary negation overflow not caught — -(i8::MIN) now fires T030.
   overflow.rs: check_op_overflow wires negation result through the
   type-range validator, mirroring the binary arithmetic path.

2. Lexer parsed decimals as i64, blocking u64 values above i64::MAX.
   Tok::Int and Expr::Int widened to i128 throughout the pipeline:
   lexer.rs, ast.rs, parser.rs, codegen.rs, cimport.rs, asm.rs, main.rs.
   Decimal and hex literals now parse via u128 then cast to i128.

Tests: 33/33 passing. Two negation tests corrected to expect T030,
one new test validates u64 decimal max (18446744073709551615).

+++ src/ast.rs
@@ -20,7 +20,7 @@ impl std::fmt::Display for Span {
 
 #[derive(Debug, Clone)]
 pub enum Expr {
-    Int(i64),
+    Int(i128),
     Str(String),
     Bool(bool),
     Var(String),
@@ -133,5 +133,5 @@ pub struct Program {
 #[derive(Debug, Clone)]
 pub struct EnumDecl {
     pub name:     String,
-    pub variants: Vec<(String, i64)>,
+    pub variants: Vec<(String, i128)>,
 }
